requirements.txt
torch==1.12.0
torchvision==0.13.0
gradio 
httpcore<0.16.0
httpx==0.23.0
datasets
huggingface-hub>=0.19
hf-transfer>=0.1.4
protobuf<4
click<8.1
pydantic~=1.0
transformers
python-dotenv